1. 支持前一交易日振幅的过滤 2. 以10倍的间隔加速或减速

+++ trunk/PlaybackConfig.h
@@ -30,10 +30,10 @@ public:
 
 	BOOL bDayOfWeek[7];
 
-	int nGapPercentage;
+	float nGapPercentage;
 	
-	int nLastDayFluctuationAbove;
+	float nLastDayFluctuationAbove;
 
-	int nLastDayFluctuationBelow;
+	float nLastDayFluctuationBelow;
 
 };
